massive addition of code from DoW
It was easiest to add the code from the DoW repo here, instead of using the fork of the repo. I will cite the original repo heavily since some of the key code there is implemented here for my own purposes. The next steps are to make sure the code runs here and places the output in the proper subdirectories. If that works, then I can finally start trying to add bequests motives to the model