edits to creating a research proposal with preliminary results
Keep this repo called old. It has a lot in it. Both regarding paper and code. But there may be a pivot regarding how the model is solved using HARK tools. If so, things will need to get migrated to a new repo. May even abandon latex/emacs structure for Myst.