small change to make sure the code runs
This is why the repo is called "old". Wealth in the utility function agent types was a bit difficult. This inspired a revamping of the code to reproduce the results. Nonetheless, these two files will still reproduce the preliminary results discussed in presentations and research proposals up to this point.